Fix frontend code
